Added version 2.8.1664

+++ chromium/components/bookmarks/browser/bookmark_model.h
@@ -137,12 +137,12 @@ class BookmarkModel : public BookmarkUndoProvider,
   // Moves |node| to |new_parent| and inserts it at the given |index|.
   void Move(const BookmarkNode* node,
             const BookmarkNode* new_parent,
-            int index);
+            size_t index);
 
   // Inserts a copy of |node| into |new_parent| at |index|.
   void Copy(const BookmarkNode* node,
             const BookmarkNode* new_parent,
-            int index);
+            size_t index);
 
   // Returns the favicon for |node|. If the favicon has not yet been loaded,
   // a load will be triggered and the observer of the model notified when done.
@@ -194,7 +194,7 @@ class BookmarkModel : public BookmarkUndoProvider,
 
   // Adds a new folder node at the specified position.
   const BookmarkNode* AddFolder(const BookmarkNode* parent,
-                                int index,
+                                size_t index,
                                 const base::string16& title,
                                 const base::string16& nickname = {},
                                 const base::string16& description = {},
@@ -206,7 +206,7 @@ class BookmarkModel : public BookmarkUndoProvider,
   // --- VIVALDI --- changed by Daniel Sig. @ 10-02-2015
   const BookmarkNode* AddFolderWithMetaInfo(
       const BookmarkNode* parent,
-      int index,
+      size_t index,
       const base::string16& title,
       const BookmarkNode::MetaInfoMap* meta_info,
       const base::string16& nickname = {},
@@ -217,7 +217,7 @@ class BookmarkModel : public BookmarkUndoProvider,
 
     // Adds a url at the specified position.
   const BookmarkNode* AddURL(const BookmarkNode* parent,
-                             int index,
+                             size_t index,
                              const base::string16& title,
                              const GURL& url,
                              const base::string16& nickname = {},
@@ -229,7 +229,7 @@ class BookmarkModel : public BookmarkUndoProvider,
   // Adds a url with a specific creation date and meta info.
   const BookmarkNode* AddURLWithCreationTimeAndMetaInfo(
       const BookmarkNode* parent,
-      int index,
+      size_t index,
       const base::string16& title,
       const GURL& url,
       const base::Time& creation_time,
@@ -343,7 +343,7 @@ class BookmarkModel : public BookmarkUndoProvider,
 
   // BookmarkUndoProvider:
   void RestoreRemovedNode(const BookmarkNode* parent,
-                          int index,
+                          size_t index,
                           std::unique_ptr<BookmarkNode> node) override;
 
   // Notifies the observers for adding every descedent of |node|.
@@ -361,14 +361,14 @@ class BookmarkModel : public BookmarkUndoProvider,
   // Adds the |node| at |parent| in the specified |index| and notifies its
   // observers.
   BookmarkNode* AddNode(BookmarkNode* parent,
-                        int index,
+                        size_t index,
                         std::unique_ptr<BookmarkNode> node);
 
   // Adds |node| to |index_| and recursisvely invokes this for all children.
   void AddNodeToIndexRecursive(BookmarkNode* node);
 
   // Returns true if the parent and index are valid.
-  bool IsValidIndex(const BookmarkNode* parent, int index, bool allow_end);
+  bool IsValidIndex(const BookmarkNode* parent, size_t index, bool allow_end);
 
   // Notification that a favicon has finished loading. If we can decode the
   // favicon, FaviconLoaded is invoked.
